requirements.txt
requests
paho-mqtt
click
python-dotenv